Make environ symbol weak
On some platforms (e.g. FreeBSD) by default -Wl,--no-undefined is passed
to the compiler, thus when generating the util library there is an
error, because `environ` is defined only in an executable. By making the
symbol weak, we avoid the issue, while still ensuring that environ works
when rz_util is linked in an executable.